requirements.txt
python-dotenv
streamlit==1.44.1
Authlib>=1.3.2
boto3
# langchain
# langchain-aws
# langchain_community
# pdfplumber
# PyPDF2
